LOGAPI-29: Refactor App component - move navbar fully to Navbar component
