+++ src/components/common/AlertComponent.vue
