+++ src/components/personal/PersonalDataEditForm.vue
@@ -42,7 +42,7 @@
 <script>
 import {useAuthStore} from "@/stores/authStore";
 import AuthDataService from "@/auth/AuthDataService";
-import AlertComponent from "@/components/AlertComponent.vue";
+import AlertComponent from "@/components/common/AlertComponent.vue";
 
 export default {
   name: "PersonalDataEditForm",
